Reference

API

NMRlab.NMRProcessor — Type

abstract type NMRProcessor <: Function end

Abstract data type for NMR processing routines. New processing tools should be declared as subtypes of NMRProcessor.

NMRlab.Chain — Method

Chain(fs::Vararg{NMRProcessor})

create a chain of processing tools, which will be applied in order (the first in the argument list is applied first)

NMRlab.FileIO — Module

module FileIO

contains low-level code to import vendor-specific NMR data files.

NMRlab.FileIO.readBrukerFID — Method

Bruker.readBrukerFID(s::String; format=Float64)

reads a Bruker fid or ser file and returns the data contained in it as a complex array. For files generated by TopSpin 4.0 and higher, the default format (Float64) will work. For TopSpin 2.0, use format=Int32.

NMRlab.FileIO.readBrukerParameterFile — Method

readBrukerParameterFile(s::String)

reads a Bruker JCAMP-DX parameter file, and returns a dict with the values. Arrays such as D0 D1 D2 etc are returned as a Julia array. Note that the Bruker parameter D0 appears as param["D"][1] due to Julia using 1-based arrays.

NMRlab.FileIO.readJEOL — Method

function readJEOL(s::IOStream) -> (header,params,data)

reads a JEOL .jdf file from the stream s. header contains a dictionary with the file header values, params is a dictionary of parameters. Each parameter contains a tuple (scaler::Int,units,value).

data is a raw data vector, which needs to be reshaped into the correct form depending on the parameters. This is handled by a separate function.

NMRlab.FileIO.reshapeJEOL — Method

function reshapeJEOL(header,params,data<:AbstractArray)

uses the header and parameter data to reshape the data. Returns an Array with the correctly shaped data.

Not implemented yet!

This function is not yet implemented. It will rely on the new data type, and will return a valid SpectData array, including the correct axes and coordinate information.

Index

NMRlab.FileIO

NMRlab.NMRProcessor

NMRlab.Chain

NMRlab.FileIO.readBrukerFID

NMRlab.FileIO.readBrukerParameterFile

NMRlab.FileIO.readJEOL

NMRlab.FileIO.reshapeJEOL